Extract set_*_attributes, sdl_clear_texture
