+++ srcs/sdl/sdl_clear_texture.c
@@ -1,38 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   loop.c                                             :+:      :+:    :+:   */
+/*   sdl_clear_texture.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/08/17 20:49:20 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/09/05 15:27:55 by oadhesiv         ###   ########.fr       */
+/*   Created: 2021/09/05 15:23:31 by oadhesiv          #+#    #+#             */
+/*   Updated: 2021/09/05 15:24:14 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "client/main.h"
-#include "loop/loop_sdl.h"
 #include "sdl/sdl_clear_texture.h"
 
-void	invalidate_sphere_vectors(t_rtv *rtv, t_sphere sphere)
-{
-	ft_putendl_fd("[invalidate_sphere_vectors] unimplemented", 2);
-	(void)rtv;
-	(void)sphere;
-}
+#include "actually_defines.h"
+#include "sdl_defines.h"
+#include "libft.h"
 
-void	loop_redraw(t_rtv *rtv)
+void	sdl_clear_texture(t_sdl *sdl)
 {
-	if (!(rtv->flags & FLAG_REDRAW))
-		return ;
-	sdl_clear_texture(rtv->sdl);
-	rtv->flags -= FLAG_REDRAW;
-	run_cluster(rtv);
-}
-
-int	loop_hook(t_rtv *rtv)
-{
-	loop_redraw(rtv);
-	loop_sdl(rtv);
-	return (0);
+	ft_bzero(sdl->buffer, WIDTH * HEIGHT * 4);
 }
